Extend CRUD API and start the great refactor

+++ app/crud/user.py
@@ -1,7 +1,12 @@
-from sqlmodel import Session
-from app.models.models import UserCreate, UserUpdate, UserPublic, User
+from sqlmodel import Session, select
+
+from app.core.dependencies import get_password_hash
+from app.models.models import User, UserCreate, UserUpdate
+
 
 def create_user(session: Session, user_create: UserCreate) -> User:
+    user_create.password = get_password_hash(user_create.password)
+
     user = User.model_validate(user_create)
     session.add(user)
     session.commit()
@@ -12,14 +17,18 @@ def get_user(session: Session, id: int) -> User | None:
     user = session.get(User, id)
     return user
 
-def get_public_user(session: Session, id: int) -> UserPublic | None:
-    user = get_user(session, id)
+def get_user_from_username(session: Session, username: str) -> User | None:
+    user = session.exec(select(User).where(User.username == username)).one()
+    return user
+
+def update_user(session: Session, id: int, user_update: UserUpdate) -> User | None:
+    if user_update.password is not None:
+        user_update.password = get_password_hash(user_update.password)
+
+    user = session.get(User, id)
     if user is None:
         return None
-    return public_user.get_public()
 
-def update_user(session: Session, id: int, user_update: UserUpdate) -> User:
-    user = session.get(User, id)
     new_user_data = user_update.model_dump(exclude_unset=True)
     user.sqlmodel_update(new_user_data)
     session.add(user)
@@ -27,7 +36,11 @@ def update_user(session: Session, id: int, user_update: UserUpdate) -> User:
     session.refresh(user)
     return user
 
-def delete_user(session: Session, id: int) -> None:
-    user = session.get(user, id)
+def delete_user(session: Session, id: int) -> bool:
+    user = session.get(User, id)
+    if user is None:
+        return False
+
     session.delete(user)
     session.commit()
+    return True
